Port the user mode OS library to GCC

+++ src/include/os.h
@@ -632,11 +632,12 @@ struct sigaction {
 // Critical sections
 //
 
-struct critsect {
-  long count;
-  long recursion;
-  tid_t owner;
-  handle_t event;
+struct critsect
+{
+    int count;
+    long recursion;
+    tid_t owner;
+    handle_t event;
 };
 
 typedef struct critsect *critsect_t;
